Update layout, image placeholders
Updated the layout for the sections to use a set `px` breakpoint object
except for sections with image backgrounds, where the padding is then
set on the actual content container within the section.

Added base64 blur backgrounds via `nextjs/image` `blur` property for
images.

Moved the AOS library initialization to the `Layout.tsx` component
instead of the `_app.tsx` file. This way, the `useBreakpoint` Chakra-UI
hook can be used to determine the `offset` property where, we want a
smaller offset for mobile to show the animations sooner than on tablet
or desktop.